requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
Authlib==1.3.2
bump2version==1.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
Django==5.1.3
django-axes==7.0.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-ranged-response==0.2.0
django-ratelimit==4.1.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.1
dparse==0.6.4
drf-yasg==1.21.8
executing==2.1.0
Faker==33.1.0
filelock==3.12.4
idna==3.10
inflection==0.5.1
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
packaging==24.2
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
psutil==6.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safety-schemas==0.0.9
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
typer==0.14.0
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
